+++ client/src/shared/vsCodeSettings/getConfiguredTestEnvironment.ts
@@ -1,8 +1,0 @@
-import { getEnvironmentSettingsKey } from "@global_shared";
-import { workspace } from "vscode";
-
-export function getConfiguredTestEnvironment() {
-    return workspace
-        .getConfiguration()
-        .get<string>(getEnvironmentSettingsKey());
-}