+++ shared/baseModel/collection.ts
@@ -1,7 +1,11 @@
+import { basename } from "path";
 import {
+    BrunoEnvironmentFile,
     CollectionData,
     CollectionDirectory,
     CollectionItem,
+    getExtensionForBrunoFiles,
+    isEnvironmentFile,
     normalizePath,
 } from "..";
 
@@ -42,6 +46,18 @@ export class Collection<T> {
 
     public getAllStoredDataForCollection() {
         return this.testData as readonly CollectionData<T>[];
+    }
+
+    public getEnvironments() {
+        return this.getAllStoredDataForCollection()
+            .filter(({ item }) => isEnvironmentFile(item))
+            .map(({ item }) => ({
+                item: item as BrunoEnvironmentFile,
+                environmentName: basename(
+                    item.getPath(),
+                    getExtensionForBrunoFiles(),
+                ),
+            }));
     }
 
     public addItem(data: CollectionData<T>) {