+++ shared/baseModel/interfaces.ts
@@ -21,7 +21,8 @@ export type AdditionalCollectionDataProvider<T> =
 
 export interface AdditionalCollectionSimpleDataProvider<T> {
     paramType: AdditionalCollectionDataProviderType.SimpleCollectionItem;
-    callback: (item: CollectionItem) => T;
+    callback: (item: CollectionItem, isCollectionRoot: boolean) => T;
+    isAdditionalDataOutdated: (oldData: T, newData: T) => boolean;
 }
 
 export interface AdditionalCollectionComplexDataProvider<T> {