+++ shared/vsCodeSettings/interfaces.ts
@@ -1,0 +1,1 @@
+export type ConfiguredEnvironmentPerCollectionSetting = Record<string, string>;